+++ core/models/session_test.go
@@ -118,10 +118,12 @@ func TestGetWaitingSessionForContact(t *testing.T) {
 	assert.Equal(t, sessionUUID, session.UUID)
 }
 
-func TestInterruptSessionsForContactsTx(t *testing.T) {
+func TestInterruptContacts(t *testing.T) {
 	ctx, rt := testsuite.Runtime(t)
 
 	defer testsuite.Reset(t, rt, testsuite.ResetData)
+
+	oa := testdb.Org1.Load(t, rt)
 
 	session1UUID, _ := insertSessionAndRun(t, rt, testdb.Ann, models.FlowTypeMessaging, models.SessionStatusCompleted, testdb.Favorites, nil)
 	session2UUID, run2UUID := insertSessionAndRun(t, rt, testdb.Ann, models.FlowTypeVoice, models.SessionStatusWaiting, testdb.Favorites, nil)
@@ -131,9 +133,8 @@ func TestInterruptSessionsForContactsTx(t *testing.T) {
 	tx := rt.DB.MustBegin()
 
 	// noop if no contacts
-	err := models.InterruptContacts(ctx, tx, map[models.ContactID]flows.SessionStatus{})
+	err := models.InterruptContacts(ctx, tx, []*models.Contact{}, flows.SessionStatusInterrupted)
 	require.NoError(t, err)
-
 	require.NoError(t, tx.Commit())
 
 	assertSessionAndRunStatus(t, rt, session1UUID, models.SessionStatusCompleted)
@@ -141,11 +142,24 @@ func TestInterruptSessionsForContactsTx(t *testing.T) {
 	assertSessionAndRunStatus(t, rt, session3UUID, models.SessionStatusWaiting)
 	assertSessionAndRunStatus(t, rt, session4UUID, models.SessionStatusWaiting)
 
+	ann, _, _ := testdb.Ann.Load(t, rt, oa)
+	bob, _, _ := testdb.Bob.Load(t, rt, oa)
+	cat, _, _ := testdb.Cat.Load(t, rt, oa)
+
+	assert.Equal(t, session2UUID, ann.CurrentSessionUUID())
+	assert.Equal(t, session3UUID, bob.CurrentSessionUUID())
+	assert.Equal(t, session4UUID, cat.CurrentSessionUUID())
+
 	tx = rt.DB.MustBegin()
 
-	err = models.InterruptContacts(ctx, tx, map[models.ContactID]flows.SessionStatus{testdb.Ann.ID: flows.SessionStatusFailed, testdb.Bob.ID: flows.SessionStatusInterrupted})
+	err = models.InterruptContacts(ctx, tx, []*models.Contact{ann}, flows.SessionStatusFailed)
 	require.NoError(t, err)
+	require.NoError(t, tx.Commit())
 
+	tx = rt.DB.MustBegin()
+
+	err = models.InterruptContacts(ctx, tx, []*models.Contact{bob}, flows.SessionStatusInterrupted)
+	require.NoError(t, err)
 	require.NoError(t, tx.Commit())
 
 	assertSessionAndRunStatus(t, rt, session1UUID, models.SessionStatusCompleted) // wasn't waiting
@@ -159,39 +173,11 @@ func TestInterruptSessionsForContactsTx(t *testing.T) {
 	assertdb.Query(t, rt.DB, `SELECT status FROM flows_flowrun WHERE uuid = $1`, run2UUID).Columns(map[string]any{"status": "F"})
 	assertdb.Query(t, rt.DB, `SELECT status FROM flows_flowrun WHERE uuid = $1`, run3UUID).Columns(map[string]any{"status": "I"})
 	assertdb.Query(t, rt.DB, `SELECT current_session_uuid, current_flow_id FROM contacts_contact WHERE id = $1`, testdb.Ann.ID).Columns(map[string]any{"current_session_uuid": nil, "current_flow_id": nil})
-}
+	assertdb.Query(t, rt.DB, `SELECT count(*) FROM contacts_contactfire`).Returns(1)
 
-func TestInterruptSessionsForChannels(t *testing.T) {
-	ctx, rt := testsuite.Runtime(t)
-
-	defer testsuite.Reset(t, rt, testsuite.ResetData)
-
-	ann1Call := testdb.InsertCall(t, rt, testdb.Org1, testdb.TwilioChannel, testdb.Ann)
-	ann2Call := testdb.InsertCall(t, rt, testdb.Org1, testdb.TwilioChannel, testdb.Ann)
-	bobCall := testdb.InsertCall(t, rt, testdb.Org1, testdb.TwilioChannel, testdb.Bob)
-	catCall := testdb.InsertCall(t, rt, testdb.Org1, testdb.VonageChannel, testdb.Cat)
-
-	session1UUID, _ := insertSessionAndRun(t, rt, testdb.Ann, models.FlowTypeMessaging, models.SessionStatusCompleted, testdb.Favorites, ann1Call)
-	session2UUID, _ := insertSessionAndRun(t, rt, testdb.Ann, models.FlowTypeMessaging, models.SessionStatusWaiting, testdb.Favorites, ann2Call)
-	session3UUID, _ := insertSessionAndRun(t, rt, testdb.Bob, models.FlowTypeMessaging, models.SessionStatusWaiting, testdb.Favorites, bobCall)
-	session4UUID, _ := insertSessionAndRun(t, rt, testdb.Cat, models.FlowTypeMessaging, models.SessionStatusWaiting, testdb.Favorites, catCall)
-
-	rt.DB.MustExec(`UPDATE ivr_call SET session_uuid = $2 WHERE id = $1`, ann1Call.ID, session1UUID)
-	rt.DB.MustExec(`UPDATE ivr_call SET session_uuid = $2 WHERE id = $1`, ann2Call.ID, session2UUID)
-	rt.DB.MustExec(`UPDATE ivr_call SET session_uuid = $2 WHERE id = $1`, bobCall.ID, session3UUID)
-	rt.DB.MustExec(`UPDATE ivr_call SET session_uuid = $2 WHERE id = $1`, catCall.ID, session4UUID)
-
-	err := models.InterruptSessionsForChannel(ctx, rt.DB, testdb.TwilioChannel.ID)
-	require.NoError(t, err)
-
-	assertSessionAndRunStatus(t, rt, session1UUID, models.SessionStatusCompleted) // wasn't waiting
-	assertSessionAndRunStatus(t, rt, session2UUID, models.SessionStatusInterrupted)
-	assertSessionAndRunStatus(t, rt, session3UUID, models.SessionStatusInterrupted)
-	assertSessionAndRunStatus(t, rt, session4UUID, models.SessionStatusWaiting) // channel not included
-
-	// check other columns are correct on interrupted session and contact
-	assertdb.Query(t, rt.DB, `SELECT count(*) FROM flows_flowsession WHERE ended_on IS NOT NULL AND current_flow_uuid IS NULL AND uuid = $1`, session2UUID).Returns(1)
-	assertdb.Query(t, rt.DB, `SELECT current_session_uuid, current_flow_id FROM contacts_contact WHERE id = $1`, testdb.Ann.ID).Columns(map[string]any{"current_session_uuid": nil, "current_flow_id": nil})
+	assert.Equal(t, flows.SessionUUID(""), ann.CurrentSessionUUID())
+	assert.Equal(t, flows.SessionUUID(""), bob.CurrentSessionUUID())
+	assert.Equal(t, session4UUID, cat.CurrentSessionUUID())
 }
 
 func TestInterruptSessionsForFlows(t *testing.T) {
@@ -239,6 +225,9 @@ func insertSessionAndRun(t *testing.T, rt *runtime.Runtime, contact *testdb.Cont
 	if status == models.SessionStatusWaiting {
 		// mark contact as being in that flow
 		rt.DB.MustExec(`UPDATE contacts_contact SET current_session_uuid = $2, current_flow_id = $3 WHERE id = $1`, contact.ID, sessionUUID, flow.ID)
+
+		// add expiration fire
+		testdb.InsertContactFire(t, rt, testdb.Org1, contact, models.ContactFireTypeSessionExpiration, "", time.Now().Add(10*time.Minute), sessionUUID)
 	}
 
 	return sessionUUID, runUUID