+++ core/runner/interrupt.go
@@ -3,8 +3,7 @@ package runner
 import (
 	"context"
 	"fmt"
-	"maps"
-	"slices"
+	"time"
 
 	"github.com/nyaruka/goflow/flows"
 	"github.com/nyaruka/goflow/flows/events"
@@ -12,51 +11,61 @@ import (
 	"github.com/nyaruka/mailroom/runtime"
 )
 
-// Interrupt interrupts the sessions for the given contacts
-// TODO rework to share contact locking code with bulk starts?
-func Interrupt(ctx context.Context, rt *runtime.Runtime, oa *models.OrgAssets, contactIDs []models.ContactID, status flows.SessionStatus) error {
-	// load our contacts
-	scenes, err := CreateScenes(ctx, rt, oa, contactIDs, nil)
+// InterruptWithLock interrupts the waiting sessions for the given contacts
+func InterruptWithLock(ctx context.Context, rt *runtime.Runtime, oa *models.OrgAssets, contactIDs []models.ContactID, status flows.SessionStatus) (map[*flows.Contact][]flows.Event, []models.ContactID, error) {
+	scenes, skipped, unlock, err := LockAndLoad(ctx, rt, oa, contactIDs, nil, 10*time.Second)
 	if err != nil {
-		return fmt.Errorf("error creating scenes for contacts: %w", err)
+		return nil, nil, err
 	}
 
-	if err := addInterruptEvents(ctx, rt, oa, scenes, status); err != nil {
-		return fmt.Errorf("error interrupting existing sessions: %w", err)
+	defer unlock() // contacts are unlocked whatever happens
+
+	if err := addInterruptEvents(ctx, rt, oa, scenes, nil, status); err != nil {
+		return nil, nil, fmt.Errorf("error interrupting existing sessions: %w", err)
 	}
 
 	if err := BulkCommit(ctx, rt, oa, scenes); err != nil {
-		return fmt.Errorf("error committing interruption scenes: %w", err)
+		return nil, nil, fmt.Errorf("error committing interruption scenes: %w", err)
 	}
 
-	return nil
+	evts := make(map[*flows.Contact][]flows.Event, len(scenes))
+	for _, s := range scenes {
+		evts[s.Contact] = s.Events()
+	}
+
+	return evts, skipped, nil
 }
 
 // adds contact interruption to the given scenes
-func addInterruptEvents(ctx context.Context, rt *runtime.Runtime, oa *models.OrgAssets, scenes []*Scene, status flows.SessionStatus) error {
-	sessions := make(map[flows.SessionUUID]*Scene, len(scenes))
+func addInterruptEvents(ctx context.Context, rt *runtime.Runtime, oa *models.OrgAssets, scenes []*Scene, sessions map[models.ContactID]flows.SessionUUID, status flows.SessionStatus) error {
+	sessionUUIDs := make([]flows.SessionUUID, 0, len(scenes))
+	byScene := make(map[*Scene]flows.SessionUUID, len(scenes))
 	for _, s := range scenes {
-		if s.DBContact.CurrentSessionUUID() != "" {
-			sessions[s.DBContact.CurrentSessionUUID()] = s
+		waitingSession := s.DBContact.CurrentSessionUUID()
+
+		// if we have a waiting session and it matches the specified session (or no session specified), add it
+		if waitingSession != "" && (sessions == nil || sessions[s.DBContact.ID()] == waitingSession || sessions[s.DBContact.ID()] == "") {
+			sessionUUIDs = append(sessionUUIDs, waitingSession)
+			byScene[s] = waitingSession
 		}
 	}
-	if len(sessions) == 0 {
+	if len(sessionUUIDs) == 0 {
 		return nil // nothing to do
 	}
 
-	runRefs, err := models.GetActiveAndWaitingRuns(ctx, rt, slices.Collect(maps.Keys(sessions)))
+	runRefs, err := models.GetActiveAndWaitingRuns(ctx, rt, sessionUUIDs)
 	if err != nil {
 		return fmt.Errorf("error getting active runs for waiting sessions: %w", err)
 	}
 
 	for _, s := range scenes {
-		if s.DBContact.CurrentSessionUUID() != "" {
-			if err := s.AddEvent(ctx, rt, oa, newContactInterruptedEvent(status), models.NilUserID); err != nil {
+		if sessionUUID := byScene[s]; sessionUUID != "" {
+			if err := s.AddEvent(ctx, rt, oa, newContactInterruptedEvent(status), models.NilUserID, ""); err != nil {
 				return fmt.Errorf("error adding contact interrupted event: %w", err)
 			}
 
 			for _, run := range runRefs[s.DBContact.CurrentSessionUUID()] {
-				if err := s.AddEvent(ctx, rt, oa, events.NewRunEnded(run.UUID, run.Flow, flows.RunStatus(status)), models.NilUserID); err != nil {
+				if err := s.AddEvent(ctx, rt, oa, events.NewRunEnded(run.UUID, run.Flow, flows.RunStatus(status)), models.NilUserID, ""); err != nil {
 					return fmt.Errorf("error adding run ended event: %w", err)
 				}
 			}